Move `Environment` to separate module
